refactor(terminal): keep struct terminal private, add row iterator
